Fix printing of commands table in yellow color. Print it as is
This is removal of unintended code. It was caught while testing
some commands and it was found that some bad word wrapped parts of
commands in the commands table were printed in yellow rather
than bright blue as expected. The bad word wrapping was due to
width issues and the color issue was due to the yellow printing
of the table. Looks like some experimental code got into the
commits and was not noticed. Gotta be careful while experimenting
things and getting rid of them instead of committing them :sweat_smile: